Move generators to another file
